Create Level One
The first level is the maze. I changed 1.txt to make a maze instead.
I added a test in TileTypeTest to create a door tile type.
I added a test in TileColorMapTest to make the doors red.
I added a test in TilePainter to make sure the door tiles are painted.
I implemented the key movement methods using ATDD.