Add bridge movement
The player can now move on the passable bridge tiles 'B'.
The player cannot go through the not passable bridge tiles 'E'.
I refactored the code by writing movePlayerVertically() in GameEngine.
The method deals with movement on the bridge since the player can only
move up and down on the bridge.