Overhauled Fog Of War System
Significant changes to fog of war system. Instead of using LIT and UNLIT
variants for each tile, the system now uses a single UNLIT tile type.
The GameEngine will maintain a second parallel hashmap for visible
tiles, which is now used for all UI methods. Visible tiles are updated
every time the player moves. Currently the player is the only light
source with a hard-coded radius of 1 emanating from the player point
object. Also added tests to ensure that tiles become unlit when they are
not in the player light radius. Still need to do some refactoring for
the visible tile methods.