Updated Tile Color Tests
I updated NOT_PASSABLE tiles to be Color.GRAY so that UNLIT tiles could
be Color.BLACK (feels more natural) and the two would be
distinguishable, but I neglected to update the color tests. Updated the
color tests to check for the new NOT_PASSABLE color and also test for
UNLIT tile color.