Create Deactivated Door Type
I added a feature file for movements into deactivated doors which have
type Q.
This led to the creation of a test for Q tile types in TileTypeTest.
I added the deactivated door type in TileType.
I added tests in TileMapColorTest and TilePainterTest to make the
deactivated door red and paint it.
I realized I didn't need paint methods for all distinct types of tiles.
So I wrote one paint method for the doors (normal & deactivated) and one
for the bridges (passable & not passable).