Implement door deactivation
I wanted the program to be able to change a door to the deactivated type
right before adding the tile to game engine.
So, I added a door deactivation test in TileTypeTest.
I also added a test in LevelCreatorIntegrationTest to check that the
door type changed to Q.
It resulted in adding a method to LevelCreatorITHelper to create levels
with doors.
This created a problem in MovementDoor.feature for move up, so I deleted
the move up scenario and added the test in GameEngineTest instead.

I changed 2.txt so that there are 2 doors at the end of the bridge
instead of 1.
I fixed the code smells.