Implement activate door randomly
Fixed code smell.
I created RandomWrapper class and used it to test choosing a random door
to activate in LevelCreatorTest.
I used dependency injection to test the use of RandomWrapper.
I added activateRandomDoor method in LevelCreator.
I included randomness in the deactivateAllDoorsButOne method to make
sure only one door is active on level 1.
This created a problem with MovementDoor.feature, so I deleted the file
and wrote tests for movements into doors in GameEngineTest instead.
I deleted the tests to create level with doors in
LevelCreatorIntegrationTest since the doors are now randomly
deactivated.