Implement going to next level
Level is no longer final.
I added tests to check that the player goes to the next level once they
reach a door.
I tested that the game exits when the next level is invalid, i.e. there
is no next level.
I added a line of code in DungeonCrawler's runIfEnoughTimeHasElapsed()
to integrate going to next level.