Limit moves
Now, the player can only bump into a deactivated door twice and bump
into a not passable bridge three times.
This is to make the game a little more difficult.
I added tests to increment counters.
I added tests to make sure the counters are back to 0 for the next
level.