GameEngineTest.jav I added the following tests: coin_counter_increases_when_moving_to_coin() coin_tile_becomes_passable_when_once_player_gathers_it() player_set_on_coin_tile_when_moving_to_it()
GameEngine.java
Added code to pass tests to KeyRight into TyleType.COIN
Refactored movingToTile() to include code for going into TyleType.COIN
Added getCoinCount() and refactored coinCount++ into increaseCoinCount()

TilePainterTest.java & TilePainter.java
Spiked code in TilePainter to determine parameter needs, then wrote paint_text() test. Wrote code to pass it.
I went on to write the tests and code in GamePanelTest & GamePanel for showing coinCount.
The position was a bit too close to the border and appeared next to lower quadrant. Which made me come back and change my code.
I spiked to find that subtracting by 5 put it in a much more suitable position.  Wrote the test paint_text_with_vertical_offset() and added parameter verticalOffset. Then rewrote the code and added a verticalOffsetParameter to pass the test. The vertical offset as a parameter is important, in case I wanted to paint text at the bottom of the screen.
With verticalOffset in mind, I wanted to see how a 2nd counter would look horizontally so I spiked the code in GamePanel to add a 2nd coinCounter, and it wasn’t viewable.   I added 10 to the tileWidth parameter of the 2nd coinCounter and saw it slightly offset of the current one.  After a couple tries, I found that when I multiplied the tileWidth*4 to find that it was now next to the coinCounter in a viewable area.  Thought that multiplying by a horizontalPosition parameter would be useful for multiple items on the top of the screen.  Wrote the test paint_text_with_horizontal_position() and wrote the code to pass it.  Editing the parameters on my previous tests and code.

GamePaneTest.java
Rename paint() to paint_tiles()
Added paint_text(), and refactored it to match parameters as what was described previously

GamePanel.java
I wrote in a String coinCount to paint and added paintText method, and didn’t like the view, as described previously, I adjusted and eventually went with offsets that were viewable.  Added parameters to tests and then in the code.  I added a test for having the levels on the scree.  This led to adding also added  String levelCount to the method, and adjusted its offset to position 1.

TileColorMapTest.java & TileColor.java
Changed Color of OBSTACLE to RED, and made PREVIOUS_LEVEL & NEXT_LEVEL WHITE.