Create Level Three
I created the third level with 3.txt.
I added a test in TileTypeTest to create an obstacle tile type.
I added a test in TileColorMapTest to make the obstacles yellow.
I added a test in TilePainter to make sure the obstacle tiles are
painted.
I changed player_cannot_go_to_an_invalid_level() in GameEngineTest to
test that player can't go past level 3.
I changed the 2 to 3 in advanceIfValidLevel() in GameEngine. That way
the player can advance up to level 3.