Implement Obstacle destruction
Now, when the player collides three times with an obstacle tile, the
tile becomes passable.
I added MovementObstacle.feature which led me to add step definitions in
the MovementStepDefs file.
I added tests in GameEngineTest to test that the obstacle become
passable after every 3 collisions.
After that there was an error with MovementObstacle, which led me to
included changeObstacleToPassable() to all the key movements.