Add timer to level 3
I added a LevelTimer class.
I did TDD for all the methods except for runLevel() which is a spike.
I tried to test it but I could not find a way to test it correctly yet.
I added a method to SystemWrapper.
I added a test in DungeonCrawlerTest to test that the correct methods
are called in the right order to make the timer work.
The timer only works on level 3.